+++ packages/dgt-id-proxy/lib/util/verify-upstream-jwk.ts
@@ -1,6 +1,6 @@
 import { decode } from 'jose/util/base64url';
-import { throwError, from, Observable } from 'rxjs';
-import { switchMap, map } from 'rxjs/operators';
+import { throwError, from, Observable, of } from 'rxjs';
+import { switchMap, map, tap } from 'rxjs/operators';
 import { JWK, JWTVerifyResult } from 'jose/types';
 import { parseJwk } from 'jose/jwk/parse';
 import { jwtVerify } from 'jose/jwt/verify';
@@ -71,3 +71,4 @@ export const verifyUpstreamJwk = (token: string, upstreamUrl: string): Observabl
   );
 
 };
+