+++ packages/inrupt-solid-client/CHANGELOG.md
@@ -3,6 +3,87 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.7.1](https://github.com/digita-ai/dgt-id-broker/compare/v0.7.0...v0.7.1) (2021-10-14)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-client
+
+
+
+
+
+## [0.7.0](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.7...v0.7.0) (2021-10-14)
+
+
+### **Features**
+
+* port solid (sdk) service from dgt-shared ([#244](https://github.com/digita-ai/dgt-id-broker/issues/244)) ([dd20110](https://github.com/digita-ai/dgt-id-broker/commit/dd20110fe1ba387d15e9e9fe496fdabf7ebe893a))
+
+
+
+### [0.6.7](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.6...v0.6.7) (2021-09-29)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-client
+
+
+
+
+
+### [0.6.6](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.5...v0.6.6) (2021-09-29)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-client
+
+
+
+
+
+### [0.6.5](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.4...v0.6.5) (2021-09-27)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-client
+
+
+
+
+
+### [0.6.4](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.3...v0.6.4) (2021-09-27)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-client
+
+
+
+
+
+### [0.6.3](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.1...v0.6.3) (2021-09-20)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-client
+
+
+
+
+
+### [0.6.2](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.1...v0.6.2) (2021-09-16)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-client
+
+
+
+
+
+### [0.6.1](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.0...v0.6.1) (2021-09-16)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-client
+
+
+
+
+
+## [0.6.0](https://github.com/digita-ai/dgt-id-broker/compare/v0.5.0...v0.6.0) (2021-09-09)
+
+**Note:** Version bump only for package @digita-ai/inrupt-solid-client
+
+
+
+
+
 ## [0.5.0](https://github.com/digita-ai/dgt-id-broker/compare/v0.4.1...v0.5.0) (2021-09-02)
 
 **Note:** Version bump only for package @digita-ai/inrupt-solid-client