+++ providers/solid-compliant-oidc-provider/CHANGELOG.md
@@ -3,6 +3,93 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.7.1](https://github.com/digita-ai/dgt-id-broker/compare/v0.7.0...v0.7.1) (2021-10-14)
+
+**Note:** Version bump only for package @digita-ai/solid-oidc-provider
+
+
+
+
+
+## [0.7.0](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.7...v0.7.0) (2021-10-14)
+
+
+### **Features**
+
+* port solid (sdk) service from dgt-shared ([#244](https://github.com/digita-ai/dgt-id-broker/issues/244)) ([dd20110](https://github.com/digita-ai/dgt-id-broker/commit/dd20110fe1ba387d15e9e9fe496fdabf7ebe893a))
+
+
+
+### [0.6.7](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.6...v0.6.7) (2021-09-29)
+
+**Note:** Version bump only for package @digita-ai/solid-oidc-provider
+
+
+
+
+
+### [0.6.6](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.5...v0.6.6) (2021-09-29)
+
+
+### **Features**
+
+* returning-upstream-error-responses ([#226](https://github.com/digita-ai/dgt-id-broker/issues/226)) ([1435256](https://github.com/digita-ai/dgt-id-broker/commit/1435256c77b6ca8cad53a2405e05633e6006f100))
+
+
+
+### [0.6.5](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.4...v0.6.5) (2021-09-27)
+
+**Note:** Version bump only for package @digita-ai/solid-oidc-provider
+
+
+
+
+
+### [0.6.4](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.3...v0.6.4) (2021-09-27)
+
+**Note:** Version bump only for package @digita-ai/solid-oidc-provider
+
+
+
+
+
+### [0.6.3](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.1...v0.6.3) (2021-09-20)
+
+
+### **Bug Fixes**
+
+* add state to function call in authRequest ([#221](https://github.com/digita-ai/dgt-id-broker/issues/221)) ([c75fc35](https://github.com/digita-ai/dgt-id-broker/commit/c75fc35221ade38dede25e9ad00855d58163eb59))
+* jose types ([#220](https://github.com/digita-ai/dgt-id-broker/issues/220)) ([9394070](https://github.com/digita-ai/dgt-id-broker/commit/93940703179c5d41f8b0385c92bfe446fc316d09))
+
+
+
+### [0.6.2](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.1...v0.6.2) (2021-09-16)
+
+
+### **Bug Fixes**
+
+* jose types ([#220](https://github.com/digita-ai/dgt-id-broker/issues/220)) ([9394070](https://github.com/digita-ai/dgt-id-broker/commit/93940703179c5d41f8b0385c92bfe446fc316d09))
+
+
+
+### [0.6.1](https://github.com/digita-ai/dgt-id-broker/compare/v0.6.0...v0.6.1) (2021-09-16)
+
+
+### **Features**
+
+* id-kit demo ([#212](https://github.com/digita-ai/dgt-id-broker/issues/212)) ([b54a287](https://github.com/digita-ai/dgt-id-broker/commit/b54a287ce3bdaac8b1f7130ed3b68bb9e1f9e977))
+
+
+
+## [0.6.0](https://github.com/digita-ai/dgt-id-broker/compare/v0.5.0...v0.6.0) (2021-09-09)
+
+
+### **Features**
+
+* major refactor of inner workings and usage of store ([#211](https://github.com/digita-ai/dgt-id-broker/issues/211)) ([ca5ccc7](https://github.com/digita-ai/dgt-id-broker/commit/ca5ccc7997024e474682198010abd1e249873f16))
+
+
+
 ## [0.5.0](https://github.com/digita-ai/dgt-id-broker/compare/v0.4.1...v0.5.0) (2021-09-02)
 
 **Note:** Version bump only for package @digita-ai/solid-oidc-provider